Updated key binding utils.
