--- MbyronModsCommon/MbyronModsCommon/KeyBinding/KeyHelper.cs
@@ -1,10 +1,0 @@
-﻿using UnityEngine;
-
-namespace MbyronModsCommon {
-    public class KeyHelper {
-        public static bool IsControlDown() => Input.GetKey(KeyCode.LeftControl) || Input.GetKey(KeyCode.RightControl);
-        public static bool IsShiftDown() => Input.GetKey(KeyCode.LeftShift) || Input.GetKey(KeyCode.RightShift);
-        public static bool IsAltDown() => Input.GetKey(KeyCode.LeftAlt) || Input.GetKey(KeyCode.RightAlt);
-
-    }
-}
